HotelLinen Amount changed to int.Input Amout Validation done withe dialog

+++ HotelLinenManagerV2.ApplicationServices/API/Domain/Models/HotelLinen.cs
@@ -12,7 +12,7 @@ namespace HotelLinenManagerV2.ApplicationServices.API.Domain.Models
         public decimal PricePerKg { get; set; }
         public string Description { get; set; }
         public string Color { get; set; }
-        public ushort Amount { get; set; }
+        public int Amount { get; set; }
         public int CompanyId { get; set; }
         public Company Company { get; set; }
         public List<WarehauseDetail> WarehauseDetails { get; set; }
