+++ HotelLinenManagerV2.ApplicationServices/API/Domain/Models/LaundryServiceDetail.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int LaundryServiceId { get; set; }
         public int HotelLinenId { get; set; }
-        public ushort Amount { get; set; }
+        public int Amount { get; set; }
         public string HotelLinenName { get; set; }
         public string Color { get; set; }
     }
